Cleanup redis connections when websockets disconnect.
Previously the redis subscriber wouldn't be closed or return to the
connection pool on websocket disconnect.  When running a few load tests
(each eg. 1000 connections) after another, we noticed that the later
load tests had more failures due to too many file handles.

The StrictRedis would release the connection on __del__ which is
fine with low churn, as the garbage collector would eventually close
the socket.  But to pro-actively close unneeded sockets, we must
explicitly release them ourselves.

This code is my own, but I was inspired by the pull request from
https://github.com/kaipeng/django-websocket-redis